requirements.txt
# Backend dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
newspaper3k>=0.2.8
google-generativeai>=0.3.1

# Frontend dependencies
streamlit>=1.28.0
plotly>=5.18.0
